Setup application with normal light flash with a print to console
Set up the user story with the same progress we had in our artifact video from last sprint. To have a led on board turn on and off from button press. Have included a set up to include a serial output be sent to a terminal listening on a specifc baud rate depending on the current state of the application the statement will change. Have been unable to fully test the correctness of this solution as my current terminal program does not let me change the baud rate. will continue working on this solution with valid testing