US33 Task 50
Completed a protocol for understanding the values used in the software implementation of the decoder. For this protocol to make sense to a user, there should be an update to the main code in which the values of the variables discussed is also sent over USART with the corresponding messages already being sent. (newA and newB values)