US66 Task 72
Removed USART0 as it was not being used and took pins needed by TCA0. Added TCA0 to the project, as the stopwatch feature cannot be used and the watch window is not a good option for performance.  Began a text document to outline how to use the timer coutner peripherals