Task 192 finished setting up ble functionality in app
Finished the framework for ble functionality on the app. this includes the ability to search for a device, connect and disconnect to device, and search for services and characteristics (the ability to transmit data). No functionality other than the basic connection with the ble device has been implemented at this time